+++ gems/aws-sdk-core/lib/aws-sdk-core/plugins/jsonvalue_converter.rb
@@ -11,31 +11,43 @@ module Aws
 
         def call(context)
           context.operation.input.shape.members.each do |m, ref|
-            convert_jsonvalue(m, ref, context.params)
+            convert_jsonvalue(m, ref, context.params, 'params')
           end
           @handler.call(context)
         end
 
-        def convert_jsonvalue(m, ref, params)
-          return if params.nil? || params[m].nil?
+        def convert_jsonvalue(m, ref, params, context)
+          return if params.nil? || !params.key?(m)
 
           if ref['jsonvalue']
-            param_value = params[m]
-            unless param_value.respond_to?(:to_json)
-              raise ArgumentError, "The value of params[#{m}] is not JSON serializable."
-            end
-            params[m] = param_value.to_json
-          elsif ref.shape.is_a?(Seahorse::Model::Shapes::StructureShape)
-            ref.shape.members.each do |member_m, ref|
-              convert_jsonvalue(member_m, ref, params[m])
-            end
-          elsif ref.shape.is_a?(Seahorse::Model::Shapes::ListShape)
-            if ref.shape.member['jsonvalue']
-              params[m] = params[m].map do |v|
-                v.to_json
+            params[m] = serialize_jsonvalue(params[m], "#{context}[#{m}]")
+          else
+            case ref.shape
+            when Seahorse::Model::Shapes::StructureShape
+              ref.shape.members.each do |member_m, ref|
+                convert_jsonvalue(member_m, ref, params[m], "#{context}[#{m}]")
+              end
+            when Seahorse::Model::Shapes::ListShape
+              if ref.shape.member['jsonvalue']
+                params[m] = params[m].each_with_index.map do |v, i|
+                  serialize_jsonvalue(v, "#{context}[#{m}][#{i}]")
+                end
+              end
+            when Seahorse::Model::Shapes::MapShape
+              if ref.shape.value['jsonvalue']
+                params[m].each do |k, v|
+                  params[m][k] = serialize_jsonvalue(v, "#{context}[#{m}][#{k}]")
+                end
               end
             end
           end
+        end
+
+        def serialize_jsonvalue(v, context)
+          unless v.respond_to?(:to_json)
+            raise ArgumentError, "The value of #{context} is not JSON serializable."
+          end
+          v.to_json
         end
 
       end